Add files via upload
One of the finest ways to diversify net worth is to invest in land, even if it is undeveloped raw property. It is one of the biggest steps to invest in land. Taking these steps needs a lot of understanding and research. In this project, the primary goal is to identify the relationship between the locality of the house, previous land value, previous improvement value, current improvement value, and current land value to predict the future land value in 5 years of land in Vancouver City.